# Added fix for non ASCII TLDs