This is a simple fix so we could start from here.

Benchmarks:
```
benchmark                    old ns/op     new ns/op     delta
BenchmarkStrictEmpty-4       13.7          14.0          +2.19%
BenchmarkStrictSingle-4      25797         28526         +10.58%
BenchmarkStrictMany-4        81277         82469         +1.47%
BenchmarkRelaxedEmpty-4      9566          15405         +61.04%
BenchmarkRelaxedSingle-4     61348         100386        +63.63%
BenchmarkRelaxedMany-4       164582        277470        +68.59%
```

We can see here that it becomes more expensive, we need to think about it. 

Regarding #32