@mvdan 2019-10-24
I'd also like to add - this is a complex issue to solve, so please understand that this isn't a simple merge.