@mvdan 2019-10-25
I've ended up going for a simpler fix: removing the feature entirely in 15684420e2b89d5320d909e8f05655c101ab18bc. See the commit message for details. Thanks again!